# swagger-codegen/jaxrs-cxf-cdi/swagger.json
{
  "swagger" : "2.0",
  "info" : {
...
    "version" : "1.0.0",
    "title" : "Samsara API"
  },
  "host" : "api.samsara.com",
  "basePath" : "/v1",
...
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/sensors/cargo" : {
      "post" : {
        "tags" : [ "Sensors" ],
        "summary" : "/sensors/cargo",
        "description" : "Get cargo monitor status (empty / full) for requested sensors.",
        "operationId" : "get_sensors_cargo",
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "sensorParam",
          "description" : "Group ID and list of sensor IDs to query.",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "groupId", "sensors" ],
            "properties" : {
              "groupId" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 101,
                "description" : "Group ID to query."
              },
              "sensors" : {
                "type" : "array",
                "example" : [ 122 ],
                "description" : "List of sensor IDs to query.",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of sensor objects containing the current cargo status reported by each sensor.",
            "schema" : {
              "$ref" : "#/definitions/CargoResponse"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sensors/door" : {
      "post" : {
        "tags" : [ "Sensors" ],
        "summary" : "/sensors/door",
        "description" : "Get door monitor status (closed / open) for requested sensors.",
        "operationId" : "get_sensors_door",
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "sensorParam",
          "description" : "Group ID and list of sensor IDs to query.",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "groupId", "sensors" ],
            "properties" : {
              "groupId" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 101,
                "description" : "Group ID to query."
              },
              "sensors" : {
                "type" : "array",
                "example" : [ 122 ],
                "description" : "List of sensor IDs to query.",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of sensor objects containing the current door status reported by each sensor.",
            "schema" : {
              "$ref" : "#/definitions/DoorResponse"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sensors/history" : {
      "post" : {
        "tags" : [ "Sensors" ],
        "summary" : "/sensors/history",
        "description" : "Get historical data for specified sensors. This method returns a set of historical data for the specified sensors in the specified time range and at the specified time resolution.",
        "operationId" : "get_sensors_history",
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "historyParam",
          "description" : "Group ID, time range and resolution, and list of sensor ID, field pairs to query.",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "endMs", "groupId", "series", "startMs", "stepMs" ],
            "properties" : {
              "endMs" : {
                "type" : "integer",
                "example" : 1462881998034,
                "description" : "End of the time range, specified in milliseconds UNIX time."
              },
              "fillMissing" : {
                "type" : "string",
                "enum" : [ "withNull", "withPrevious" ],
                "default" : "withNull"
              },
              "groupId" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 101,
                "description" : "Group ID to query."
              },
              "series" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "description" : "Sensor ID and field to query.",
                  "properties" : {
                    "field" : {
                      "type" : "string",
                      "example" : "ambientTemperature",
                      "description" : "Field to query.",
                      "enum" : [ "ambientTemperature", "probeTemperature", "currentLoop1Raw", "currentLoop1Mapped", "currentLoop2Raw", "currentLoop2Mapped", "pmPowerTotal", "pmPhase1Power", "pmPhase2Power", "pmPhase3Power", "pmPhase1PowerFactor", "pmPhase2PowerFactor", "pmPhase3PowerFactor" ]
                    },
                    "widgetId" : {
                      "type" : "integer",
                      "format" : "int64",
                      "example" : 1,
                      "description" : "Sensor ID to query."
                    }
                  },
                  "required" : [ "field", "widgetId" ]
                }
              },
              "startMs" : {
                "type" : "integer",
                "example" : 1462878398034,
                "description" : "Beginning of the time range, specified in milliseconds UNIX time."
              },
              "stepMs" : {
                "type" : "integer",
                "example" : 3600000,
                "description" : "Time resolution for which data should be returned, in milliseconds. Specifying 3600000 will return data at hour intervals."
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of results objects, each containing a time and a datapoint for each requested sensor/field pair.",
            "schema" : {
              "$ref" : "#/definitions/SensorHistoryResponse"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sensors/humidity" : {
      "post" : {
        "tags" : [ "Sensors" ],
        "summary" : "/sensors/humidity",
        "description" : "Get humidity for requested sensors. This method returns the current relative humidity for the requested sensors.",
        "operationId" : "get_sensors_humidity",
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "sensorParam",
          "description" : "Group ID and list of sensor IDs to query.",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "groupId", "sensors" ],
            "properties" : {
              "groupId" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 101,
                "description" : "Group ID to query."
              },
              "sensors" : {
                "type" : "array",
                "example" : [ 122 ],
                "description" : "List of sensor IDs to query.",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of sensor objects containing the current humidity reported by each sensor.",
            "schema" : {
              "$ref" : "#/definitions/HumidityResponse"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sensors/list" : {
      "post" : {
        "tags" : [ "Sensors" ],
        "summary" : "/sensors/list",
        "description" : "Get sensor objects. This method returns a list of the sensor objects in the Samsara Cloud and information about them.",
        "operationId" : "get_sensors",
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "groupParam",
          "description" : "Group ID to query.",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "groupId" ],
            "properties" : {
              "groupId" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 101,
                "description" : "Group ID to query."
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of sensor objects.",
            "schema" : {
              "type" : "object",
              "properties" : {
                "sensors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Sensor"
                  }
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sensors/temperature" : {
      "post" : {
        "tags" : [ "Sensors" ],
        "summary" : "/sensors/temperature",
        "description" : "Get temperature for requested sensors. This method returns the current ambient temperature (and probe temperature if applicable) for the requested sensors.",
        "operationId" : "get_sensors_temperature",
        "parameters" : [ {
          "name" : "access_token",
          "in" : "query",
          "description" : "Samsara API access token.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "sensorParam",
          "description" : "Group ID and list of sensor IDs to query.",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "groupId", "sensors" ],
            "properties" : {
              "groupId" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 101,
                "description" : "Group ID to query."
              },
              "sensors" : {
                "type" : "array",
                "example" : [ 122 ],
                "description" : "List of sensor IDs to query.",
                "items" : {
                  "type" : "integer",
                  "format" : "int64"
                }
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of sensor objects containing the current temperature reported by each sensor.",
            "schema" : {
              "$ref" : "#/definitions/TemperatureResponse"
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "CargoResponse" : {
      "type" : "object",
      "properties" : {
        "groupId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 101
        },
        "sensors" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "properties" : {
              "cargoEmpty" : {
                "type" : "boolean",
                "example" : true,
                "description" : "Flag indicating whether the current cargo is empty or loaded."
              },
              "id" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 122,
                "description" : "ID of the sensor."
              },
              "name" : {
                "type" : "string",
                "example" : "Trailer Cargo Sensor",
                "description" : "Name of the sensor."
              }
            }
          }
        }
      },
      "description" : "Contains the current cargo status of a sensor."
    },
...
    "DoorResponse" : {
      "type" : "object",
      "properties" : {
        "groupId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 101
        },
        "sensors" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "properties" : {
              "doorClosed" : {
                "type" : "boolean",
                "example" : true,
                "description" : "Flag indicating whether the current door is closed or open."
              },
              "id" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 122,
                "description" : "ID of the sensor."
              },
              "name" : {
                "type" : "string",
                "example" : "Trailer Door Sensor",
                "description" : "Name of the sensor."
              }
            }
          }
        }
      },
      "description" : "Contains the current door status of a sensor."
    },
...
    "ErrorResponse" : {
      "type" : "string",
      "description" : "Error message describing why the request failed."
    },
...
    "HumidityResponse" : {
      "type" : "object",
      "properties" : {
        "groupId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 101
        },
        "sensors" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "properties" : {
              "humidity" : {
                "type" : "integer",
                "example" : 53,
                "description" : "Currently reported relative humidity in percent, from 0-100."
              },
              "id" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 122,
                "description" : "ID of the sensor."
              },
              "name" : {
                "type" : "string",
                "example" : "Freezer Humidity Sensor",
                "description" : "Name of the sensor."
              }
            }
          }
        }
      },
      "description" : "Contains the current humidity of a sensor."
    },
...
    "Sensor" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 123,
          "description" : "ID of the sensor."
        },
        "macAddress" : {
          "type" : "string",
          "example" : "11:11:11:11:11:11",
          "description" : "MAC address of the sensor."
        },
        "name" : {
          "type" : "string",
          "example" : "Freezer ABC",
          "description" : "Name of the sensor."
        }
      },
      "description" : "Contains information about a sensor."
    },
    "SensorHistoryResponse" : {
      "type" : "object",
      "properties" : {
        "results" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "properties" : {
              "series" : {
                "type" : "array",
                "description" : "List of datapoints, one for each requested (sensor, field) pair.",
                "items" : {
                  "type" : "integer",
                  "format" : "int64",
                  "example" : 1
                }
              },
              "timeMs" : {
                "type" : "integer",
                "example" : 1453449599999,
                "description" : "Timestamp in UNIX milliseconds."
              }
            }
          }
        }
      },
      "description" : "Contains the results for a sensor history request. Each result contains a timestamp and datapoint for each requested (sensor, field) pair."
    },
...
    "TemperatureResponse" : {
      "type" : "object",
      "properties" : {
        "groupId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 101
        },
        "sensors" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "properties" : {
              "ambientTemperature" : {
                "type" : "integer",
                "example" : 11057,
                "description" : "Currently reported ambient temperature in millidegrees celsius."
              },
              "id" : {
                "type" : "integer",
                "format" : "int64",
                "example" : 122,
                "description" : "ID of the sensor."
              },
              "name" : {
                "type" : "string",
                "example" : "Freezer Temp Sensor",
                "description" : "Name of the sensor."
              },
              "probeTemperature" : {
                "type" : "integer",
                "example" : -20145,
                "description" : "Currently reported probe temperature in millidegrees celsius. If no probe is connected, this parameter will not be reported."
              }
            }
          }
        }
      },
      "description" : "Contains the current temperatures of a sensor."
    },
...
  },
...
}